import-service/requirements.txt
ics==0.7.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-api-python-client==2.31.0